gui/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
pydantic==2.7.1
dtaidistance==2.4.0
wfdb==4.1.2
numpy==1.26.4
python-multipart==0.0.9
